Remove dependency on BiomaRt server; use local files instead.
Add config file that specifies path to large local files that will be
queried in lieu of public BiomaRt server to retrieve sequences for
Ensembl transcripts. If not on the Broad servers, will need to download
these files yourself as they live in my Broad directory and change the
paths in the config file. Config file contains info for GRCh37 and
GRCh38 even though code currenty only uses GRCh37, so that once
Oncotator changes I will be equipped to switch to GRCh38
build/annotation.